CQM no longer uses verify_job interface.  Interface removed.
It was determined that the interface to verify jobs was no longer
needed, and instead a hold check should be added to set_jobs and
add_jobs.  This allowed us to simplify the interface.  The check code
no longer raises top level exceptions, and now handles partially
successful lists better.