Fix for ticket #4317: Job Modification and walltime
This fixes a race where a job can be modified during the scheduling loop
and be scheduled with a different time than the queue-manager has.  This
can cause some very strange artifacts in the job log, and can cause the
database to have inaccurate walltimes, along with a user thinking a
job's walltime was modified when it, in effect, wasn't.  This causes a
job that has a mismatch in its walltime with the one the scheduler has
to not start up.  The job should start on the next scheduler pass with
the correct walltime.